refactor: move Permissao para pasta Enums e remove rota de teste
